% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishinitseasonEst.r
\name{fishinitseasonEst}
\alias{fishinitseasonEst}
\title{fishinitseasonEst function
Initial week of the proposed season}
\usage{
fishinitseasonEst(
  dataset,
  Year,
  Week,
  RegionName,
  Strata,
  StrataVolume,
  RegionVolume,
  VolumeSample,
  N_individual,
  first_thr = 0.05,
  durationEst = 8
)
}
\arguments{
\item{dataset}{A dataset contains the input data.
This dataset must include the following variables.}

\item{Year}{A variable of Year of data collection.}

\item{Week}{A variable of calender week of data collection.}

\item{RegionName}{A variable of names of regions where data were collected.}

\item{Strata}{A variable of strata within the corresponding \code{RegionName} where data were collected.}

\item{StrataVolume}{A variable of volume of the corresponding \code{Strata} where data were collected.
Depending on the sampling design, it could be area of a strata.}

\item{RegionVolume}{A variable of volume of the corresponding region where data were collected.
Depending on the sampling design, it could be area of a region.}

\item{VolumeSample}{A variable of volume of water sampled.
Depending on the sampling design, it could be time of tow duration or area swept.}

\item{N_individual}{A variable of number of individual sampled.}

\item{first_thr}{A numeric value in between 0 and 1.
It is used to determine the first week of the year in which
the cumulative weekly density estimates surpass \code{first_thr}*100\% of
the total densities observed across all weeks of sampling.}

\item{durationEst}{Integer (>0).
The estimated duration of weeks the season will last.
For example, if the season is estimated to end 8 weeks after
the estimated first week (see \code{first_thr}), then the input should be 8.}
}
\value{
This function returns a dataframe which includes a variable of year and
a variable of estimated initial week of the proposed season
}
\description{
fishinitseasonEst function
Initial week of the proposed season
}
\details{
The data were assumed collected from a fishery-independent survey following
a stratified random sampling design with i regions within the sampling area and
s strata within a given region. The survey was conducted over week for each year.
The initial week of the proposed season is estimated as the first week of the year when
the cumulative weekly density estimates surpass \code{first_thr}*100\% of
the total densities observed across all weeks of sampling.
}
\examples{
\dontrun{
 seasonQ<-fishinitseasonEst(dataset=fishQ,Year="yr",
                                Week="wk",
                                RegionName="regname",
                                Strata="stra",
                                StrataVolume="stravol",
                                RegionVolume="regvol",
                                VolumeSample="volsamp",
                                N_individual="nind",
                                first_thr=0.05,durationEst=7
 )
 seasonQ
 
}
 
}
