% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{fishX}
\alias{fishX}
\title{Hypothetical fish survey data for species X}
\format{
An object of class \code{data.frame} with 4452 rows and 10 columns.
}
\usage{
fishX
}
\description{
Hypothetical fish survey data for species X
}
\examples{
\dontrun{
 fishX
}
}
\keyword{datasets}
