% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fishai.r
\name{plot.fishai.default}
\alias{plot.fishai.default}
\title{plot.fishai function
Plot annual abundance index}
\usage{
\method{plot}{fishai.default}(df, ..., Year = NULL, AbundIndex = NULL)
}
\arguments{
\item{df}{A dataset contains a variable of year and a variable of abundanc indices.}

\item{...}{further arguments passed to or from other methods.}

\item{Year}{A variable of Year of data collection.}

\item{AbundIndex}{A variable of annual abundance indices.}
}
\value{
This function returns a graph of the annual abundance indices time series.
}
\description{
plot.fishai function
Plot annual abundance index
}
\examples{
\dontrun{
 aiQ<-fishai(dataset=fishQ,Year="yr",
            Week="wk",
            RegionName="regname",
            Strata="stra",
            StrataVolume="stravol",
            RegionVolume="regvol",
            VolumeSample="volsamp",
            N_individual="nind",
            SeasonEst=TRUE,first_thr=0.05,durationEst=7)
 plot.fishai(aiQ,Year=Year,AbundIndex=AbundIndex)

 aiX<-fishai(dataset=fishX,Year="yr",
            Week="wk",
            RegionName="regname",
            Strata="stra",
            StrataVolume="stravol",
            RegionVolume="regvol",
            VolumeSample="volsamp",
            N_individual="nind",
            SeasonEst=FALSE)
 plot.fishai(aiX,Year=Year,AbundIndex=AbundIndex)

}

}
\concept{plot.fishai methods}
