% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fishai.r
\name{plot.fishai}
\alias{plot.fishai}
\title{plot.fishai function
Plot annual abundance index}
\usage{
plot.fishai(df, ..., Year = NULL, AbundIndex = NULL)
}
\arguments{
\item{df}{A dataset contains a variable of year and a variable of abundanc indices.}

\item{...}{further arguments passed to or from other methods.}

\item{Year}{A variable of Year of data collection.}

\item{AbundIndex}{A variable of annual abundance indices.}
}
\description{
This the generic melt function. See the following functions
}
