% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fishstd.r
\name{plot.fishstd}
\alias{plot.fishstd}
\title{plot.fishstd function
Plot spatiotemporal distribution for each region and each year}
\usage{
\method{plot}{fishstd}(df, legend.title = NULL, plot.title = NULL)
}
\arguments{
\item{df}{A dataset contains density for each region and each year.
Each row is a region, and each column is a year.}

\item{legend.title}{Title for the color key.}

\item{plot.title}{Title for the graph.}
}
\value{
This function returns a heatmap showing spatiotemporal distribution over regions and years.
}
\description{
plot.fishstd function
Plot spatiotemporal distribution for each region and each year
}
\examples{
\dontrun{
 stdQ<-fishstd(dataset=fishQ,Year="yr",
              Week="wk",
              RegionName="regname",
              Strata="stra",
              StrataVolume="stravol",
              RegionVolume="regvol",
              VolumeSample="volsamp",
              N_individual="nind",
              SeasonEst=TRUE,first_thr=0.05,durationEst=7
 )
 plot.fishstd(df=stdQ,legend.title="Standardized Density",plot.title="SpeciesQ")

 stdX<-fishstd(dataset=fishX,Year="yr",
              Week="wk",
              RegionName="regname",
              Strata="stra",
              StrataVolume="stravol",
              RegionVolume="regvol",
              VolumeSample="volsamp",
              N_individual="nind",
              SeasonEst=FALSE
 )
 plot.fishstd(df=stdX,legend.title="Standardized Density",plot.title="SpeciesX")

}

}
