% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.initseason.r
\name{plot.initseason}
\alias{plot.initseason}
\title{plot.initseason function
Plot initial weeks}
\usage{
\method{plot}{initseason}(df, Year = NULL, first_wk = NULL)
}
\arguments{
\item{df}{A dataset contains a variable of year and a variable of initial week.}

\item{Year}{A variable of Year of data collection.}

\item{first_wk}{A variable of estimated first week of the season.}
}
\value{
This function returns a graph of the initial week time series.
}
\description{
plot.initseason function
Plot initial weeks
}
\examples{
\dontrun{
 seasonQ<-fishinitseasonEst(dataset=fishQ,Year="yr",
                                Week="wk",
                                RegionName="regname",
                                Strata="stra",
                                StrataVolume="stravol",
                                RegionVolume="regvol",
                                VolumeSample="volsamp",
                                N_individual="nind",
                                first_thr=0.05,durationEst=7
 )
 plot.initseason(seasonQ,Year=Year,first_wk=first_wk)

 }
}
