% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfishai.r
\name{plotfishai}
\alias{plotfishai}
\title{plotfishai function
Plot annual abundance index#'}
\usage{
plotfishai(df, ..., Year = NULL, AbundIndex = NULL, se = TRUE)
}
\arguments{
\item{df}{A dataset contains a variable of year and a variable of abundance indices.}

\item{...}{arguments passed to other methods}

\item{Year}{A variable of Year of data collection.}

\item{AbundIndex}{A variable of annual abundance indices.}

\item{se}{Logical. Standard error will be plotted out if TRUE.
the default is TRUE.}
}
\value{
This function returns a graph of the abundance indices time series.
}
\description{
plotfishai function
Plot annual abundance index#'
}
\examples{
\dontrun{
 aiQ<-fishai(dataset=fishQ,Year="yr",
            Week="wk",
            RegionName="regname",
            Strata="stra",
            StrataVolume="stravol",
            RegionVolume="regvol",
            VolumeSample="volsamp",
            N_individual="nind",
            SeasonEst=TRUE,first_thr=0.05,durationEst=7)
 plotfishai(aiQ)

 aiX<-fishai(dataset=fishX,Year="yr",
            Week="wk",
            RegionName="regname",
            Strata="stra",
            StrataVolume="stravol",
            RegionVolume="regvol",
            VolumeSample="volsamp",
            N_individual="nind",
            SeasonEst=FALSE)
 plotfishai(aiX,Year=Year,AbundIndex=AbundIndex)

}

}
